pyproject.toml
[project]
name = "wiki-stream-analytics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "google-auth>=2.48.0",
    "google-auth-httplib2>=0.3.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.1",
    "testcontainers[kafka]>=4.14.1",
    "ty>=0.0.15",
    "watchfiles>=1.1.1",
]
processing = [
    "apache-flink==2.0.1",
    "requests>=2.32.5",
]
producer = [
    "aiohttp>=3.13.3",
    "aiokafka>=0.13.0",
]

[tool.uv.extra-build-dependencies]
pyarrow = ["setuptools<81"]
grpcio-tools = ["setuptools<81"]

[tool.pytest.ini_options]
markers = [
    "unit: Unit tests (fast, no external services).",
    "integration: Integration tests (Kafka, Docker, etc.).",
]
